New solution, types, wrong answer.
Adds newPos that takes a state and distance and gives a new coordinate.
findBaseWithCoords now uses a Seq.fold instead of let rec.
Updated examples to new types instead of using strings.

Gives 303, which is the wrong answer, but answers correctly for the examples. Uses an active pattern to match headings as NESW instead of just a tuple.